refactor(GameFactory): write interface & rename Genorator to Factory
Rename Genorator.java -> CarRacingFactory.java
Write new interface
