+++ src/main/java/racingcar/factory/CarRacingFactory.java
@@ -1,16 +1,16 @@
-package racingcar.generator;
+package racingcar.factory;
 
 import racingcar.model.Car;
 import racingcar.model.Game;
 import racingcar.model.Player;
 import racingcar.model.RacingGame;
 
-public class Genorator {
-    static Player getPlayer(String name) {
+public class CarRacingFactory implements GameFactory{
+    public Player getPlayer(String name) {
         return new Car(name);
     }
 
-    static Game getGamer() {
+    public Game getGame() {
         return new RacingGame();
     }
 }
