+++ src/main/java/racingcar/factory/GameFactory.java
@@ -1,0 +1,10 @@
+package racingcar.factory;
+
+import racingcar.model.Game;
+import racingcar.model.Player;
+
+public interface GameFactory {
+    Player getPlayer(String name);
+
+    Game getGame();
+}
